move tests to separate dir to avoid problems with caching on ci
