+++ pkg/tests/sample/suites/sample_suite_test.go
@@ -1,6 +1,4 @@
-// +build suite
-
-package sample
+package suites
 
 import (
 	"testing"
